#168 refactor: 공부방 생성 모달 컴포넌트 분리
